cleanup import Data and Axis
