+++ src/pymodaq/utils/exceptions.py
@@ -20,10 +20,6 @@ class ViewerError(Exception):
     pass
 
 
-class DataSourceError(Exception):
-    pass
-
-
 class ExpectedError(Exception):
     """Raised in the tests made for pymodaq plugins"""
     pass
